requirements.txt
torch
pandas
tqdm
Pillow
numpy
torchvision
kornia
thop
ptflops
pytorch_wavelets
matplotlib 
optuna
nevergrad